# Conflicts:
#	EndWar/Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset

+++ EndWar/Assets/Scripts/Lee/Player/PlayerCtrl.cs
@@ -6,7 +6,7 @@ using Photon.Pun;
 public class PlayerCtrl : MonoBehaviour
 {
     [PunRPC]
-    public void GetDamage(int damage)
+    public void GetDamage(float damage)
     {
         Debug.Log("아프다~~~~~ 행~복~~해~~~줘~~어~~~~" + this.gameObject.name + ", " + damage);
     }
